move definitions of structs
