+++ pallas-traverse/src/header.rs
@@ -1,11 +1,19 @@
 use std::borrow::Cow;
 use std::ops::Deref;
 
-use pallas_codec::minicbor;
+use pallas_codec::{minicbor, utils::KeepRaw};
 use pallas_crypto::hash::{Hash, Hasher};
 use pallas_primitives::{alonzo, babbage, byron};
 
-use crate::{wellknown::GenesisValues, Era, Error, MultiEraHeader, OriginalHash};
+use crate::{wellknown::GenesisValues, Era, Error, OriginalHash};
+
+#[derive(Debug)]
+pub enum MultiEraHeader<'b> {
+    EpochBoundary(Cow<'b, KeepRaw<'b, byron::EbbHead>>),
+    ShelleyCompatible(Cow<'b, KeepRaw<'b, alonzo::Header>>),
+    BabbageCompatible(Cow<'b, KeepRaw<'b, babbage::Header>>),
+    Byron(Cow<'b, KeepRaw<'b, byron::BlockHead>>),
+}
 
 impl<'b> MultiEraHeader<'b> {
     pub fn decode(tag: u8, subtag: Option<u8>, cbor: &'b [u8]) -> Result<Self, Error> {
