+++ pallas-traverse/src/lib.rs
@@ -1,194 +1,48 @@
 //! Utilities to traverse over multi-era block data
 
-use pallas_codec::utils::NonZeroInt;
-use pallas_codec::utils::PositiveCoin;
-use std::{borrow::Cow, fmt::Display, hash::Hash as StdHash};
+use std::fmt::Display;
 
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
-
-use pallas_codec::utils::{KeepRaw, KeyValuePairs, NonEmptyKeyValuePairs};
-use pallas_crypto::hash::Hash;
-use pallas_primitives::{alonzo, babbage, byron, conway};
 
 mod support;
 
-pub mod assets;
-pub mod auxiliary;
-pub mod block;
-pub mod cert;
-pub mod era;
+mod assets;
+mod auxiliary;
+mod block;
+mod cert;
+mod era;
 pub mod fees;
 pub mod hashes;
-pub mod header;
-pub mod input;
-pub mod meta;
-pub mod output;
+mod header;
+mod input;
+mod meta;
+mod output;
 pub mod probe;
-pub mod redeemers;
-pub mod signers;
-pub mod size;
+mod redeemers;
+mod signers;
+mod size;
 pub mod time;
-pub mod tx;
+mod tx;
 pub mod update;
-pub mod withdrawals;
-pub mod witnesses;
+mod withdrawals;
+mod witnesses;
+
+pub use assets::{MultiEraAsset, MultiEraPolicyAssets};
+pub use block::MultiEraBlock;
+pub use cert::MultiEraCert;
+pub use era::{Era, Feature};
+pub use header::MultiEraHeader;
+pub use input::{OutputRef, MultiEraInput};
+pub use meta::MultiEraMeta;
+pub use output::MultiEraOutput;
+pub use redeemers::MultiEraRedeemer;
+pub use signers::MultiEraSigners;
+pub use tx::MultiEraTx;
+pub use update::MultiEraUpdate;
+pub use withdrawals::MultiEraWithdrawals;
 
 // TODO: move to genesis crate
 pub mod wellknown;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
-#[non_exhaustive]
-pub enum Era {
-    Byron,
-    Shelley,
-    Allegra, // time-locks
-    Mary,    // multi-assets
-    Alonzo,  // smart-contracts
-    Babbage, // CIP-31/32/33
-    Conway,  // governance CIP-1694
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-#[non_exhaustive]
-pub enum Feature {
-    TimeLocks,
-    MultiAssets,
-    Staking,
-    SmartContracts,
-    CIP31,
-    CIP32,
-    CIP33,
-    CIP1694,
-}
-
-#[derive(Debug)]
-pub enum MultiEraHeader<'b> {
-    EpochBoundary(Cow<'b, KeepRaw<'b, byron::EbbHead>>),
-    ShelleyCompatible(Cow<'b, KeepRaw<'b, alonzo::Header>>),
-    BabbageCompatible(Cow<'b, KeepRaw<'b, babbage::Header>>),
-    Byron(Cow<'b, KeepRaw<'b, byron::BlockHead>>),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraBlock<'b> {
-    EpochBoundary(Box<byron::MintedEbBlock<'b>>),
-    AlonzoCompatible(Box<alonzo::MintedBlock<'b>>, Era),
-    Babbage(Box<babbage::MintedBlock<'b>>),
-    Byron(Box<byron::MintedBlock<'b>>),
-    Conway(Box<conway::MintedBlock<'b>>),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraTx<'b> {
-    AlonzoCompatible(Box<Cow<'b, alonzo::MintedTx<'b>>>, Era),
-    Babbage(Box<Cow<'b, babbage::MintedTx<'b>>>),
-    Byron(Box<Cow<'b, byron::MintedTxPayload<'b>>>),
-    Conway(Box<Cow<'b, conway::MintedTx<'b>>>),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum MultiEraOutput<'b> {
-    AlonzoCompatible(Box<Cow<'b, alonzo::TransactionOutput>>, Era),
-    Babbage(Box<Cow<'b, babbage::MintedTransactionOutput<'b>>>),
-    Conway(Box<Cow<'b, conway::MintedTransactionOutput<'b>>>),
-    Byron(Box<Cow<'b, byron::TxOut>>),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, StdHash)]
-#[non_exhaustive]
-pub enum MultiEraInput<'b> {
-    Byron(Box<Cow<'b, byron::TxIn>>),
-    AlonzoCompatible(Box<Cow<'b, alonzo::TransactionInput>>),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraCert<'b> {
-    NotApplicable,
-    AlonzoCompatible(Box<Cow<'b, alonzo::Certificate>>),
-    Conway(Box<Cow<'b, conway::Certificate>>),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraRedeemer<'b> {
-    AlonzoCompatible(Box<Cow<'b, alonzo::Redeemer>>),
-    Conway(
-        Box<Cow<'b, conway::RedeemersKey>>,
-        Box<Cow<'b, conway::RedeemersValue>>,
-    ),
-}
-
-#[derive(Debug, Clone, Default)]
-#[non_exhaustive]
-pub enum MultiEraMeta<'b> {
-    #[default]
-    Empty,
-    NotApplicable,
-    AlonzoCompatible(&'b alonzo::Metadata),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraPolicyAssets<'b> {
-    AlonzoCompatibleMint(
-        &'b alonzo::PolicyId,
-        &'b KeyValuePairs<alonzo::AssetName, i64>,
-    ),
-    AlonzoCompatibleOutput(
-        &'b alonzo::PolicyId,
-        &'b KeyValuePairs<alonzo::AssetName, u64>,
-    ),
-    ConwayMint(
-        &'b alonzo::PolicyId,
-        &'b NonEmptyKeyValuePairs<alonzo::AssetName, NonZeroInt>,
-    ),
-    ConwayOutput(
-        &'b alonzo::PolicyId,
-        &'b NonEmptyKeyValuePairs<alonzo::AssetName, PositiveCoin>,
-    ),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraAsset<'b> {
-    AlonzoCompatibleOutput(&'b alonzo::PolicyId, &'b alonzo::AssetName, u64),
-    AlonzoCompatibleMint(&'b alonzo::PolicyId, &'b alonzo::AssetName, i64),
-    ConwayOutput(&'b alonzo::PolicyId, &'b alonzo::AssetName, PositiveCoin),
-    ConwayMint(&'b alonzo::PolicyId, &'b alonzo::AssetName, NonZeroInt),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraWithdrawals<'b> {
-    NotApplicable,
-    Empty,
-    AlonzoCompatible(&'b alonzo::Withdrawals),
-    Conway(&'b conway::Withdrawals),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraUpdate<'b> {
-    Byron(u64, Box<Cow<'b, byron::UpProp>>),
-    AlonzoCompatible(Box<Cow<'b, alonzo::Update>>),
-    Babbage(Box<Cow<'b, babbage::Update>>),
-}
-
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub enum MultiEraSigners<'b> {
-    NotApplicable,
-    Empty,
-    AlonzoCompatible(&'b alonzo::RequiredSigners),
-}
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq)]
-pub struct OutputRef(Hash<32>, u64);
 
 #[derive(Debug, Error)]
 pub enum Error {
