+++ pallas-traverse/src/update.rs
@@ -50,7 +50,15 @@ pub type ExUnits = alonzo::ExUnits;
 pub type CostMdls = alonzo::CostMdls;
 pub type ProtocolVersion = alonzo::ProtocolVersion;
 
-use crate::{Era, MultiEraUpdate};
+use crate::Era;
+
+#[derive(Debug, Clone)]
+#[non_exhaustive]
+pub enum MultiEraUpdate<'b> {
+    Byron(u64, Box<Cow<'b, byron::UpProp>>),
+    AlonzoCompatible(Box<Cow<'b, alonzo::Update>>),
+    Babbage(Box<Cow<'b, babbage::Update>>),
+}
 
 impl<'b> MultiEraUpdate<'b> {
     pub fn decode_for_era(era: Era, cbor: &'b [u8]) -> Result<Self, minicbor::decode::Error> {
